shellcheck the modules/nixup/*.sh script
Fixes potential issues with escaping and others

+++ nixup/modules/nixup/nixup-resources.sh
@@ -3,10 +3,10 @@
 rmEmptyDirectories () {
 	BASE=$1
 	DIR=$2
-	if [ "$DIR" != "." -a -d "$BASE/$DIR" ]; then
-		if [ ! "$(@coreutils@/bin/ls -A $BASE/$DIR)" ]; then
-			@coreutils@/bin/rm -rf $BASE/$DIR
-			rmEmptyDirectories $BASE `@coreutils@/bin/dirname $DIR`
+	if [ "$DIR" != "." ] && [ -d "$BASE/$DIR" ]; then
+		if [ ! "$(@coreutils@/bin/ls -A "$BASE/$DIR")" ]; then
+			@coreutils@/bin/rm -rf "${BASE:?}/$DIR"
+			rmEmptyDirectories "$BASE" "$(@coreutils@/bin/dirname "$DIR")"
 		fi
 	fi
 }
@@ -16,9 +16,9 @@ unlink () {
 	FROM=$1
 	LINK=$2
 	if [ -L "$FROM/$LINK" ]; then
-		rm -r "$FROM/$LINK"
+		rm -r "${FROM:?}/$LINK"
 	fi
-	rmEmptyDirectories $FROM `@coreutils@/bin/dirname $LINK`
+	rmEmptyDirectories "$FROM" "$(@coreutils@/bin/dirname "$LINK")"
 }
 
 
@@ -28,11 +28,11 @@ link (){
 	LINK=$3
 
 	target="$TO/$LINK"
-	mkdir -p $(@coreutils@/bin/dirname $target)
-	if @coreutils@/bin/test -e $target -a \! -L $target; then
-		@coreutils@/bin/mv $target $target.backup
+	mkdir -p "$(@coreutils@/bin/dirname "$target")"
+	if [ -e "$target" ] && ! [ -L "$target" ]; then
+		@coreutils@/bin/mv "$target" "$target.backup"
 	fi
-	@coreutils@/bin/ln -sf "$FROM/$LINK" $target
+	@coreutils@/bin/ln -sf "$FROM/$LINK" "$target"
 }
 
 
@@ -40,17 +40,17 @@ linkHome () {
 	home=$1
 	resources=$2
 	resourceTarget=$3
-	for l in `@coreutils@/bin/cat $resourceTarget`; do
-		link "$home" "$resources" $l
-	done
+	while IFS= read -r line; do
+		link "$home" "$resources" "$line"
+	done < "$resourceTarget"
 }
 
 unlinkHome () {
 	home=$1
 	resourceTarget=$2
-	for l in `@coreutils@/bin/cat $resourceTarget`; do
-		unlink "$home" $l
-	done
+	while IFS= read -r line; do
+		unlink "$home" "$line"
+	done < "$resourceTarget"
 }
 
 switchHome () {
@@ -58,12 +58,14 @@ switchHome () {
 	oldResourceTargets=$2
 	newResources=$3
 	newResourceTargets=$4
-	for l in `@coreutils@/bin/comm -23 $oldResourceTargets $newResourceTargets`; do
-		unlink $home $l
-	done
-	for l in `@coreutils@/bin/comm -13 $oldResourceTargets $newResourceTargets`; do
-		link $home $newResources $l
-	done
+	@coreutils@/bin/comm -23 "$oldResourceTargets" "$newResourceTargets" | \
+		while IFS= read -r line; do
+			unlink "$home" "$line"
+		done
+	@coreutils@/bin/comm -13 "$oldResourceTargets" "$newResourceTargets" | \
+		while IFS= read -r line; do
+			link "$home" "$newResources" "$line"
+		done
 }
 
 
@@ -110,15 +112,16 @@ done
 if [ -z "$action" ]; then showSyntax; fi
 
 
-if [ "$action" == "--link" ]; then
-	linkHome $home $nixup_dir/resources $nixup_dir/resourceTargets
+if [ "$action" = "--link" ]; then
+	linkHome "$home" "$nixup_dir/resources" "$nixup_dir/resourceTargets"
 fi
 
-if [ "$action" == "--unlink" ]; then
-	unlinkHome $home $nixup_dir/resourceTargets
+if [ "$action" = "--unlink" ]; then
+	unlinkHome "$home" "$nixup_dir/resourceTargets"
 fi
 
-if [ "$action" == "--switch" ]; then
-	switchHome $home $old_nixup_dir/resourceTargets $new_nixup_dir/resources $new_nixup_dir/resourceTargets
+if [ "$action" = "--switch" ]; then
+	switchHome "$home" "$old_nixup_dir/resourceTargets" \
+		"$new_nixup_dir/resources" "$new_nixup_dir/resourceTargets"
 fi
